+++ algonaut_transaction/src/lib.rs
@@ -4,6 +4,7 @@ pub mod auction;
 pub mod builder;
 pub mod contract_account;
 pub mod error;
+pub mod signer;
 pub mod transaction;
 pub mod tx_group;
 pub mod url;
@@ -13,4 +14,5 @@ pub use builder::{
     CreateApplication, CreateAsset, DeleteApplication, DestroyAsset, FreezeAsset, OptInApplication,
     Pay, RegisterKey, TransferAsset, UpdateApplication, UpdateAsset,
 };
+pub use signer::{MultisigSigner, Signer};
 pub use transaction::{SignedTransaction, Transaction, TransactionType};